+++ typescript/vrotsc/e2e/expect/saga/actions/tasks/testSagaContext.js
@@ -1,0 +1,10 @@
+/**
+ * @return {Any}
+ */
+(function () {
+    var exports = {};
+    exports.default = function (context) {
+        context.conditionalVar = true;
+    };
+    return exports;
+});